cambios

+++ models/pago.py
@@ -5,7 +5,7 @@ class Pago(db.Model):
 
     id = db.Column(db.Integer, primary_key=True)
     tipo_pago= db.Column(db.String(50))
-    cantidad = db.Column(db.Integer(50))
+    cantidad = db.Column(db.Integer)
 
     def __init__ (self, tipo_pago, cantidad) :
         self.tipo_pago = tipo_pago
